+++ app/android/app/src/main/kotlin/oss/yniantongtian/uncloud/uncloud/MainActivity.kt
@@ -1,0 +1,5 @@
+package oss.yniantongtian.uncloud.uncloud
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()